Splash,Loading Screen with resource manage and  game restart done

+++ core/src/np/com/samundrakc/game/misc/MessageBox.java
@@ -138,7 +138,7 @@ public class MessageBox {
     }
 
 
-    public MessageBox autoHide(final int seconds, final OnOkButtonClicked onOkButtonClicked) {
+    public MessageBox autoHide(final float seconds, final OnOkButtonClicked onOkButtonClicked) {
         final Timer timer = new Timer();
         final int[] count = {0};
         timer.scheduleAtFixedRate(new TimerTask() {
